pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "de-framework"
version = "0.1.0"
description = "Data Engineering data distillation pipeline execution framework"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.7.0",
    "tenacity>=8.3.0",
    "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
datahub = ["datahub-ingestion>=0.13.0"]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
de-run = "framework.bootstrap:main"

[tool.setuptools.packages.find]
include = ["framework*", "pipelines*", "npu_service*", "code_download*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
